Python3 fixes for structure volume calculation
When the CalculateVolume function in dvhdata.py was called, there was a
call to the itervalues() method, which is not used in Python 3. This was
changed to use the itervalues method in the six library.  Additionally,
the dictionary for each contour slice now has the key 'data' instead of
the previous 'contourData' key.