Update dicomgui.py
6 missing 'break' statements added. This repairs some strange features importing data. In my case it would never find the structured data (RTSS), while I knew it was there.
These 6 bug fixes assume that there is always exactly one and one match.When this one match is found it applies a 'break' statement to the 'for' loop. However, I am not sure if there could be more than one match in the DICOM header data (more entries in a sequence?) 

So please check if this does not jeopardize how the code should work... for my particular data cases it is an improvement.